pointing_mode: virtual keys
adding a pointing_device_mode that uses the pointing_device input (the raw data) to derive virtual key states, which in turn press/hold/release key events according to a layer synchronized `pointing_virtkey_map`.

with the later the PM_MODE becomes "orthogonal" to the keyboard layer; e.g. the pointing device can be set into `PM_VIRTKEY` and after that all inputs of that device are transformed into virtual key states, that trigger key actions in correlation to the current layer.